根据新的数据库模型，进行重构

+++ src/main/java/cn/edu/szu/Bajie/entity/DishUrl.java
@@ -25,7 +25,7 @@ public class DishUrl implements Serializable {
      * 菜品id
      */
     @TableField(value = "dish_id")
-    private Integer dishId;
+    private Long dishId;
 
     /**
      * 菜品图片url
